+++ examples/06_math_ops.cu
@@ -1,17 +1,17 @@
 // 06_math_ops.cu: Arithmetic, bitwise operations, and comparison logic
 
-fn is_even(n: i64) -> bool {
+fn is_even(n: i32) -> bool {
     return (n % 2) == 0
 }
 
-fn min(a: i64, b: i64) -> i64 {
+fn min(a: i32, b: i32) -> i32 {
     if a < b {
         return a
     }
     return b
 }
 
-fn max(a: i64, b: i64) -> i64 {
+fn max(a: i32, b: i32) -> i32 {
     if a > b {
         return a
     }
